+++ pgx-utils/src/sql_entity_graph/pg_extern/argument.rs
@@ -10,7 +10,7 @@ use syn::{
 
 /// A parsed `#[pg_extern]` argument.
 ///
-/// It is created during [`PgExtern`](pgx_utils::sql_entity_graph::PgExtern) parsing.
+/// It is created during [`PgExtern`](crate::sql_entity_graph::PgExtern) parsing.
 #[derive(Debug, Clone)]
 pub struct Argument {
     pat: syn::Ident,
