+++ pgx/src/datum/mod.rs
@@ -39,7 +39,7 @@ pub use item_pointer_data::*;
 pub use json::*;
 pub use numeric::*;
 use once_cell::sync::Lazy;
-pub use sql_entity_graph::{Aggregate, RustSqlMapping};
+pub use sql_entity_graph::RustSqlMapping;
 use std::any::TypeId;
 pub use time_stamp::*;
 pub use time_stamp_with_timezone::*;
