--- pgx/src/datum/sql_entity_graph/aggregate/finalize_modify.rs
@@ -1,20 +1,0 @@
-use crate::sql_entity_graph::{PgxSql, ToSql};
-
-/// Corresponds to the `FINALFUNC_MODIFY` and `MFINALFUNC_MODIFY` in [`CREATE AGGREGATE`](https://www.postgresql.org/docs/current/sql-createaggregate.html).
-#[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, PartialOrd, Ord)]
-pub enum FinalizeModify {
-    ReadOnly,
-    Shareable,
-    ReadWrite,
-}
-
-impl ToSql for FinalizeModify {
-    fn to_sql(&self, _context: &PgxSql) -> eyre::Result<String> {
-        let value = match self {
-            FinalizeModify::ReadOnly => String::from("READ_ONLY"),
-            FinalizeModify::Shareable => String::from("SHAREABLE"),
-            FinalizeModify::ReadWrite => String::from("READ_WRITE"),
-        };
-        Ok(value)
-    }
-}
