--- pgx/src/datum/sql_entity_graph/aggregate/parallel_option.rs
@@ -1,20 +1,0 @@
-use crate::sql_entity_graph::{PgxSql, ToSql};
-
-/// Corresponds to the `PARALLEL` and `MFINALFUNC_MODIFY` in [`CREATE AGGREGATE`](https://www.postgresql.org/docs/current/sql-createaggregate.html).
-#[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, PartialOrd, Ord)]
-pub enum ParallelOption {
-    Safe,
-    Restricted,
-    Unsafe,
-}
-
-impl ToSql for ParallelOption {
-    fn to_sql(&self, _context: &PgxSql) -> eyre::Result<String> {
-        let value = match self {
-            ParallelOption::Safe => String::from("SAFE"),
-            ParallelOption::Restricted => String::from("RESTRICTED"),
-            ParallelOption::Unsafe => String::from("UNSAFE"),
-        };
-        Ok(value)
-    }
-}
